+++ LaunchDarkly/LaunchDarkly/Models/ConnectionInformation.swift
@@ -7,9 +7,6 @@
 //
 
 import Foundation
-#if canImport(SystemConfiguration)
-import SystemConfiguration
-#endif
 
 public struct ConnectionInformation: Codable {
     public enum ConnectionMode: String, Codable {
@@ -44,7 +41,6 @@ public struct ConnectionInformation: Codable {
     }
     
     public struct Constants {
-        static let noInterval: TimeInterval = 0
         static let noCode: Int = 0
         static let unknownError: String = "Unknown Error"
         static let decodeError: String =  "Unable to Decode error."
@@ -96,14 +92,10 @@ public struct ConnectionInformation: Codable {
     }
     
     //Restores ConnectionInformation from UserDefaults if it exists
-    static func uncacheConnectionInformation(config: LDConfig, ldClient: LDClient, connectionInformationStore: ConnectionInformationStore) -> ConnectionInformation {
-        if var maybeConnectionInformation = connectionInformationStore.retrieveStoredConnectionInformation() {
-            Log.debug("Retrieved ConnectionInformation from cache")
-            maybeConnectionInformation.currentConnectionMode = connectionModeCheck(config: config, ldClient: ldClient)
-            return maybeConnectionInformation
-        } else {
-            return ConnectionInformation(currentConnectionMode: ConnectionInformation.connectionModeCheck(config: config, ldClient: ldClient), lastConnectionFailureReason: .none)
-        }
+    static func uncacheConnectionInformation(config: LDConfig, ldClient: LDClient, connectionInformationStore: ConnectionInformationStore, clientServiceFactory: ClientServiceCreating) -> ConnectionInformation {
+        var connectionInformation = connectionInformationStore.retrieveStoredConnectionInformation() ?? clientServiceFactory.makeConnectionInformation()
+        connectionInformation.currentConnectionMode = connectionModeCheck(config: config, ldClient: ldClient)
+        return connectionInformation
     }
     
     //Used for updating lastSuccessfulConnection when connected to streaming
@@ -114,45 +106,17 @@ public struct ConnectionInformation: Codable {
         return connectionInformation
     }
     
-    #if canImport(SystemConfiguration)
-    //Sourced from: https://stackoverflow.com/a/39782859
-    static func isConnectedToNetwork() -> Bool {
-        var zeroAddress = sockaddr_in(sin_len: 0, sin_family: 0, sin_port: 0, sin_addr: in_addr(s_addr: 0), sin_zero: (0, 0, 0, 0, 0, 0, 0, 0))
-        zeroAddress.sin_len = UInt8(MemoryLayout.size(ofValue: zeroAddress))
-        zeroAddress.sin_family = sa_family_t(AF_INET)
-        
-        let defaultRouteReachability = withUnsafePointer(to: &zeroAddress) {
-            $0.withMemoryRebound(to: sockaddr.self, capacity: 1) {zeroSockAddress in
-                SCNetworkReachabilityCreateWithAddress(nil, zeroSockAddress)
-            }
-        }
-        
-        var flags: SCNetworkReachabilityFlags = SCNetworkReachabilityFlags(rawValue: 0)
-        if SCNetworkReachabilityGetFlags(defaultRouteReachability!, &flags) == false {
-            return false
-        }
-        
-        let isReachable = (flags.rawValue & UInt32(kSCNetworkFlagsReachable)) != 0
-        let needsConnection = (flags.rawValue & UInt32(kSCNetworkFlagsConnectionRequired)) != 0
-        let reachability = (isReachable && !needsConnection)
-        
-        return reachability
-    }
-    #endif
-    
     //Used for updating ConnectionInformation inside of LDClient.setOnline
     static func onlineSetCheck(flagSynchronizer: LDFlagSynchronizing, connectionInformation: inout ConnectionInformation, ldClient: LDClient, config: LDConfig) -> ConnectionInformation {
         var connectionInformationVar = ConnectionInformation.lastSuccessfulConnectionCheck(flagSynchronizer: flagSynchronizer, connectionInformation: &connectionInformation)
-        if ldClient.isOnline {
+        
+        if !NetworkReporter.isConnectedToNetwork() {
+            connectionInformationVar.currentConnectionMode = ConnectionInformation.ConnectionMode.offline
+        } else if ldClient.isOnline {
             connectionInformationVar.currentConnectionMode = effectiveStreamingMode(runMode: ldClient.runMode, config: config, ldClient: ldClient) == LDStreamingMode.streaming ? ConnectionInformation.ConnectionMode.streaming : ConnectionInformation.ConnectionMode.polling
         } else {
             connectionInformationVar.currentConnectionMode = ConnectionInformation.ConnectionMode.offline
         }
-        #if canImport(SystemConfiguration)
-        if !isConnectedToNetwork() {
-            connectionInformationVar.currentConnectionMode = ConnectionInformation.ConnectionMode.offline
-        }
-        #endif
         return connectionInformationVar
     }
     
@@ -216,10 +180,8 @@ public struct ConnectionInformation: Codable {
     internal static func backgroundBehavior(connectionInformation: inout ConnectionInformation, runMode: LDClientRunMode, config: LDConfig, ldClient: LDClient) -> ConnectionInformation {
         if ConnectionInformation.effectiveStreamingMode(runMode: runMode, config: config, ldClient: ldClient) == LDStreamingMode.streaming && runMode == .background {
             connectionInformation.lastSuccessfulConnection = Date().timeIntervalSince1970
-            connectionInformation.currentConnectionMode = .polling
-        } else {
-            connectionInformation.currentConnectionMode = .polling
         }
+        connectionInformation.currentConnectionMode = .polling
         return connectionInformation
     }
 }
@@ -263,7 +225,7 @@ extension ConnectionInformation.LastConnectionFailureReason {
     }
 }
 
-extension TimeInterval {
+private extension TimeInterval {
     func stringFromTimeInterval() -> String {
         let time = NSInteger(self)
         let milli = Int((self.truncatingRemainder(dividingBy: 1)) * 1000)
