+++ LaunchDarkly/LaunchDarkly/Service Objects/NetworkReporter.swift
@@ -1,0 +1,44 @@
+//
+//  File.swift
+//  LaunchDarkly
+//
+//  Created by Joe Cieslik on 8/27/19.
+//  Copyright © 2019 Catamorphic Co. All rights reserved.
+//
+
+import Foundation
+#if canImport(SystemConfiguration)
+import SystemConfiguration
+#endif
+
+class NetworkReporter {
+    #if canImport(SystemConfiguration)
+    //Sourced from: https://stackoverflow.com/a/39782859
+    static func isConnectedToNetwork() -> Bool {
+        var zeroAddress = sockaddr_in(sin_len: 0, sin_family: 0, sin_port: 0, sin_addr: in_addr(s_addr: 0), sin_zero: (0, 0, 0, 0, 0, 0, 0, 0))
+        zeroAddress.sin_len = UInt8(MemoryLayout.size(ofValue: zeroAddress))
+        zeroAddress.sin_family = sa_family_t(AF_INET)
+        
+        let defaultRouteReachability = withUnsafePointer(to: &zeroAddress) {
+            $0.withMemoryRebound(to: sockaddr.self, capacity: 1) {zeroSockAddress in
+                SCNetworkReachabilityCreateWithAddress(nil, zeroSockAddress)
+            }
+        }
+        
+        var flags: SCNetworkReachabilityFlags = SCNetworkReachabilityFlags(rawValue: 0)
+        if SCNetworkReachabilityGetFlags(defaultRouteReachability!, &flags) == false {
+            return false
+        }
+        
+        let isReachable = (flags.rawValue & UInt32(kSCNetworkFlagsReachable)) != 0
+        let needsConnection = (flags.rawValue & UInt32(kSCNetworkFlagsConnectionRequired)) != 0
+        let reachability = (isReachable && !needsConnection)
+        
+        return reachability
+    }
+    #else
+    static func isConnectecToNetwork() -> Bool {
+        return true
+    }
+    #endif
+}
